Three pages created
I have created three pages. 1-For uploading the main story 2-For uploading the episodes 3- For managing the upload of both main story and episodes. The problem I have here is that When I upload a main story I cant go to the upload episodes. I know I need to have a button  to go to the upload episodes. Can I get help how to think and go through it and also how can I be able to save the episodes with a variable or property that can link it to the main story? Because I know I need to have them linked so later when I have a page for people to buy them I should have an if statement which says if an episodes has this main story id then it belongs to that main story. I hope you understand what I mean.