Log in by email and not username
I was having a problem where by everytime I try to log in I was getting an invalid response. I went through some materials and I understood that finding user through username was a problem and therefore the best way was through email since in the log in page it requires username to be email and not just username.