User Role has been updated to a defaultm mode
I made two changes. First in the ApplicationUser where I assigned the role to be customer so that when the user fills his or her profile data it will be automatically the role is a customer. Second I I went to change code in the profile page, instead of giving options to choose I set the role as a read only. So this User Role will take the value automatically from the APPLICATION USER CLASS