If statement was added to show before and after profile filling
I added an if statement to make sure when the user fills his or her data right after saving it should show the profile in another style. I also learned that the format was not correct so I discovered a single way to format the page and it was to press ⌥ Option + ⌘ Command + L. This is the reason it seems like some block of codes were completely removed and others added. but only few codes were added and the rest of the visible changed is because of formating