LIST of users page created
I created a page to load users and also to be able to assign people's roles. It is working As I expected.

NB. I full understand the code for inserting password or a secret work is not necessary but I just included it because I think its cool I did a nice job to create that block of code ;)