#10
I created codes to pull the name of the user so that on the welcome page the user will get to be welcomed by their names. This will help to make the user feel valued and not just like a random website visitor. An if statement was used to implement this